Range Simplification
I wrote some simple code in what was the boundQuery function (and is now the standardizeQuery function) to simplify common range phrases, such as "between * and *", "from * to *", and "*-*".

I also had to make a small edit to the tokenization process to properly catch ranges, which sometimes are expressed without spaces (fx $20-30).

+++ src/database.py
@@ -38,7 +38,7 @@ jobs = Table("Jobs", 0, [1],
     ["title", "TEXT NOT NULL"],
     ["salary", "TEXT", ['$']], # this may be problematic since the salary is in a string bucket
     ["description", "TEXT"],
-    ["rating", "NUMERIC"], # [0,5], -1
+    ["rating", "INTEGER"], # [0,5], -1
     ["company", "TEXT"],
     ["location", "TEXT"],
     ["hq", "TEXT"],
